infrastructure/requirements.txt
aws-cdk-lib==2.127.0
constructs==10.3.0
boto3==1.28.57
